Restructure certain tests
Move CVLayer tests to noise_modes\test_cv_noise ; SCZ and invert_permutation tests to cv\test_cv_ops ; macronode reduction tests to test_cv_noise. Delete tests_macro_reduce and reduce tests and imports in test_graphstates accordingly.
