+++ tests/qubit_codes/test_graphstates.py
@@ -16,21 +16,13 @@
 # pylint: disable=protected-access
 
 import string
-import logging
-
-from datetime import datetime
 
 import networkx as nx
 import numpy as np
 import numpy.random as rand
-from numpy.random import default_rng as rng
 import pytest
-import scipy.sparse as sp
 
-from flamingpy.codes import SurfaceCode
 from flamingpy.codes.graphs import EGraph
-from flamingpy.cv.ops import SCZ_mat, SCZ_apply
-from flamingpy.noise import CVLayer
 
 # A NetworkX random graph of size N for use in this module.
 N = 20
@@ -85,206 +77,3 @@ class TestEGraph:
 
     # def test_slice_coords(self):
     # pass
-
-
-class TestCVHelpers:
-    """Tests for CVLayer helper functions."""
-
-    @pytest.mark.parametrize(
-        "sparse, expected_out_type", sorted([(True, sp.coo_matrix), (False, np.ndarray)])
-    )
-    def test_SCZ_mat_sparse_param(self, random_graph, sparse, expected_out_type):
-        """Tests the SCZ_mat function outputs sparse or dense arrays."""
-        SCZ = SCZ_mat(random_graph[2], sparse=sparse)
-        assert isinstance(SCZ, expected_out_type)
-
-    def test_SCZ_mat(self, random_graph):
-        """Tests the SCZ_mat function."""
-        SCZ = SCZ_mat(random_graph[1])
-        SCZ_sparse = SCZ_mat(random_graph[2])
-        # Check if SCZ_mat adjusts type of output matrix based on
-        # type of input.
-        assert isinstance(SCZ, np.ndarray)
-        assert isinstance(SCZ_sparse, sp.coo_matrix)
-        # Check that structure of SCZ matrix is correct.
-        for mat in (SCZ, SCZ_sparse.toarray()):
-            assert np.array_equal(mat[:N, :N], np.identity(N))
-            assert np.array_equal(mat[:N, N:], np.zeros((N, N)))
-            assert np.array_equal(mat[N:, :N], random_graph[1])
-            assert np.array_equal(mat[N:, N:], np.identity(N))
-
-    @pytest.mark.parametrize("one_shot", sorted([True, False]))
-    @pytest.mark.parametrize("n", sorted([1, 2]))
-    def test_SCZ_apply(self, random_graph, one_shot, n):
-        """Test SCZ matrix application."""
-
-        adj = random_graph[1]
-        SCZ = SCZ_mat(adj)
-        N = adj.shape[0]
-        quads_shape = [N * 2] * n
-        quads = np.random.rand(*quads_shape)
-
-        if n == 1:
-            expected_quads = SCZ_mat(adj).dot(quads)
-        else:
-            expected_quads = SCZ.dot(SCZ.dot(quads).T).T
-
-        new_quads = SCZ_apply(adj, quads, one_shot=one_shot)
-
-        assert np.allclose(new_quads, expected_quads)
-
-
-now = datetime.now()
-int_time = int(str(now.year) + str(now.month) + str(now.day) + str(now.hour) + str(now.minute))
-logging.info("the following seed was used for random number generation: %i", int_time)
-
-
-class TestCVLayer:
-    """Tests for functions in the CVLayer class."""
-
-    def test_empty_init(self, random_graph):
-        """Test the instantiation of CVLayer from codes and EGraphs."""
-        G = CVLayer(random_graph[0], delta=0.1, states=None)
-        G_array = nx.to_numpy_array(G.egraph)
-        H = CVLayer(EGraph(random_graph[0]), delta=0.1, states=None)
-        H_array = nx.to_numpy_array(H.egraph)
-        # Check that the _N attribute is populated with the number
-        # of nodes in the random graph.
-        assert G._N == len(random_graph[0])
-        # Check that instantiating a CVLayer with an EGraph or with
-        # a regular NetworkX graph has the same effect.
-        assert np.array_equal(random_graph[1], H_array)
-        assert np.array_equal(random_graph[1], G_array)
-        CVRHG = CVLayer(SurfaceCode(2), delta=0.1)
-        assert CVRHG._N == len(SurfaceCode(2).graph)
-
-    def test_all_GKP_init(self, random_graph):
-        """Test the all-GKP initialization of EGraph."""
-        G = CVLayer(random_graph[0], delta=0.1)
-        G.populate_states()
-        n = len(random_graph[0])
-        for node in G.egraph:
-            assert G.egraph.nodes[node]["state"] == "GKP"
-        assert len(G.states["p"]) == 0
-        assert len(G.p_inds) == 0
-        assert np.array_equal(G.states["GKP"], np.arange(n))
-        assert np.array_equal(G.gkp_inds, np.arange(n))
-
-    @pytest.mark.parametrize("p_swap", sorted([0, 0.99 * rng(int_time).random() + 0.01, 1]))
-    def test_hybridize(self, random_graph, p_swap):
-        """Test whether CVLayer properly populates p-squeezed states for non-
-        zero p-swap."""
-        n = len(random_graph[0])
-        # Test all-p case
-        G = CVLayer(random_graph[0], delta=0.1, p_swap=1)
-        G.populate_states()
-        for node in G.egraph:
-            assert G.egraph.nodes[node]["state"] == "p"
-        assert len(G.states["GKP"]) == 0
-        assert np.array_equal(G.states["p"], np.arange(n))
-        # Test various swap-out probabilities; mean p population
-        # should be close to p_swap parameter, within tolerance.
-        p_list = []
-        for _ in range(1000):
-            G = CVLayer(random_graph[0], delta=0.1, p_swap=p_swap)
-            G.populate_states()
-            p_list += [len(G.states["p"]) / n]
-        p_prob = sum(p_list) / 1000
-        assert np.isclose(p_prob, p_swap, rtol=1e-1)
-
-    def test_state_indices(self, random_graph):
-        """Test that _states, p_inds, and gkp_inds are populated with the
-        correct indices."""
-        n = len(random_graph[0])
-        num_ps = rng().integers(n)
-        p_inds = rng().choice(n, num_ps, replace=False)
-        gkp_inds = list(set(np.arange(n)) - set(p_inds))
-        G = CVLayer(random_graph[0], delta=0.1, states={"p": p_inds})
-        G.populate_states()
-        assert np.array_equal(G.states.get("p"), p_inds)
-        assert np.array_equal(G.states.get("GKP"), gkp_inds)
-        assert np.array_equal(G.p_inds, p_inds)
-        assert np.array_equal(G.gkp_inds, gkp_inds)
-
-    @pytest.mark.parametrize("order", sorted(["initial", "two-step"]))
-    def test_apply_noise(self, random_graph, order):
-        """Check delta, sampling_order attributes with default noise model."""
-        delta = rng().random()
-        G = CVLayer(random_graph[0], delta=delta, sampling_order=order)
-        assert G.delta == delta
-        assert G._sampling_order == order
-
-    def test_grn_model_init(self, random_graph):
-        """Compare expected noise objects (quadratures, covariance matrix) with
-        those obtained through the GRN model in CVLayer with the initial
-        sampling order."""
-        delta = rng().random()
-
-        n = len(random_graph[0])
-        G = CVLayer(random_graph[0], delta=delta)
-        H = CVLayer(random_graph[0], delta=delta, p_swap=1)
-        G.populate_states(), H.populate_states()
-        H._covs_sampler()
-        G._covs_sampler()
-        init_noise_all_GKP = np.full(2 * n, (delta / 2) ** 0.5, dtype=np.float32)
-        init_noise_all_p = np.array(
-            [1 / (2 * delta) ** 0.5] * n + [(delta / 2) ** 0.5] * n, dtype=np.float32
-        )
-        assert np.array_equal(G._init_covs, init_noise_all_GKP)
-        assert np.array_equal(H._init_covs, init_noise_all_p)
-
-        # G.apply_noise(sampling_order="final")
-        # H.apply_noise(sampling_order="final")
-        # noise_cov_all_GKP = SCZ_apply(G._adj, np.diag(init_noise_all_GKP) ** 2)
-        # noise_cov_all_p = SCZ_apply(H._adj, np.diag(init_noise_all_p) ** 2)
-        # assert np.array_equal(G.Noise_cov.toarray(), noise_cov_all_GKP)
-        # assert np.array_equal(H.Noise_cov.toarray(), noise_cov_all_p)
-
-    def test_grn_model_two_step(self, random_graph):
-        """Compare expected noise objects (quadratures, covariance matrix) with
-        those obtained through the GRN model in CVLayer with the final sampling
-        order."""
-        delta = rng().random()
-
-        n = len(random_graph[0])
-        G = CVLayer(random_graph[0], delta=delta, sampling_order="two-step")
-        H = CVLayer(random_graph[0], delta=delta, p_swap=1, sampling_order="two-step")
-        G.populate_states(), H.populate_states()
-        G._means_sampler()
-        H._means_sampler()
-        assert np.max(H._init_means[:n]) <= 2 * np.sqrt(np.pi)
-        assert np.min(H._init_means[:n]) >= 0
-        assert np.isclose(np.max(G._init_means[:n]), np.sqrt(np.pi))
-        assert np.isclose(np.min(G._init_means[:n]), 0)
-
-    @pytest.mark.parametrize("order", sorted(["initial"]))
-    def test_measure_hom(self, random_graph, order):
-        """Test closeness of average homodyne outcomes value to 0 in the all-
-        GKP high-squeezing limit."""
-        delta = 0.0001
-        n = len(random_graph[0])
-        G = CVLayer(random_graph[0], delta=delta)
-        G.populate_states()
-        G.measure_hom("p")
-        G.measure_hom("q")
-        outcomes_p = []
-        outcomes_q = []
-        for node in G.egraph:
-            outcomes_p += [G.egraph.nodes[node]["hom_val_p"]]
-            outcomes_q += [G.egraph.nodes[node]["hom_val_q"]]
-        assert np.isclose(sum(outcomes_p) / n, 0, atol=1e-1)
-        assert np.isclose(sum(outcomes_q) / n, 0, atol=1e-1)
-        # TODO: test two-step sampling
-
-    # def test_eval_Z_probs(self, random_graph):
-    #     """Test that p_phase and p_phase_cond attribute get populated when
-    #     phase error probabilities are evaluated."""
-    #     pass
-    #     G = CVLayer(random_graph[0])
-    #     G.apply_noise(noise_model(delta=rng().random(), order="final"))
-    #     G.measure_hom("p")
-    #     G.eval_Z_probs()
-    #     G.eval_Z_probs(cond=True)
-    #     for node in G.egraph:
-    #         assert "p_phase" in G.egraph.nodes[node]
-    #         assert "p_phase_cond" in G.egraph.nodes[node]
